requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.12.5
google-genai>=0.1.0